Update_User
Update_User

+++ AIops_BackEnd/src/main/java/com/aiops_web/dto/UserPermissionDTO.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 public class UserPermissionDTO extends User implements Serializable {
-    private List<Integer> permissions;
+    private List<Long> permissions;
 
     private String token;
 
@@ -31,12 +31,12 @@ public class UserPermissionDTO extends User implements Serializable {
      * @param permitStr
      * @return
      */
-    private List<Integer> permissionToList(String permitStr) {
-        List<Integer> list = new ArrayList<>();
+    private List<Long> permissionToList(String permitStr) {
+        List<Long> list = new ArrayList<>();
         String[] strArr = permitStr.split("[|]");
         for (int i = 0; i < strArr.length; i++) {
             if (strArr[i].equals("1")) {
-                list.add(i+1);
+                list.add(i + 1L);
             }
         }
